requirements.txt
fastapi
uvicorn
python-multipart
pillow
jinja2